V1 cart creation and order creation from cart

+++ src/main/java/com/railswad/deliveryservice/dto/CartItemDTO.java
@@ -2,12 +2,10 @@ package com.railswad.deliveryservice.dto;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class CartItemDTO {
     private Long itemId;
     private Integer quantity;
-    private BigDecimal unitPrice;
+    private Double unitPrice;
     private String specialInstructions;
 }
